+++ src/components/audio-embed/styles.tsx
@@ -1,0 +1,6 @@
+import { type Module } from "@divi/types";
+import { type AudioEmbedAttrs } from "./types";
+
+export const ModuleStyles = (_props: Module.StylesProps<AudioEmbedAttrs>) => {
+  return null;
+};