+++ src/components/audio-embed/types.ts
@@ -1,0 +1,14 @@
+import { type ModuleLibrary } from "@divi/types";
+
+export interface AudioEmbedAttrs {
+  module: object;
+  audioBaseUrl: object;
+}
+
+export type AudioEmbedEditProps =
+  ModuleLibrary.Module.Component.EditProps<AudioEmbedAttrs>;
+
+export interface AudioEmbedAppProps {
+  slug: string;
+  audioBaseUrl: string;
+}