+++ src/index.ts
@@ -12,6 +12,7 @@ import { relatedItems } from "./components/related-items";
 import { campaignProgress } from "./components/campaign-progress";
 import { campaignDonate } from "./components/campaign-donate";
 import { ctaBox } from "./components/cta-box";
+import { audioEmbed } from "./components/audio-embed";
 
 // Import icons.
 import "./module-icons";
@@ -97,5 +98,14 @@ addAction(
   "vvp/ctaBox",
   () => {
     registerModule(ctaBox.metadata, omit(ctaBox, "metadata"));
+  },
+);
+
+// Register Audio Embed module with DIVI.
+addAction(
+  "divi.moduleLibrary.registerModuleLibraryStore.after",
+  "vvp/audioEmbed",
+  () => {
+    registerModule(audioEmbed.metadata, omit(audioEmbed, "metadata"));
   },
 );